add .dotenv to playground
