+++ apps/vite-playground/package.json
@@ -3,23 +3,24 @@
     "@esbuild-plugins/node-globals-polyfill": "0.2.3",
     "@esbuild-plugins/node-modules-polyfill": "0.2.2",
     "@thorswap-lib/evm-web3-wallets": "workspace:*",
+    "@thorswap-lib/keepkey": "workspace:*",
     "@thorswap-lib/keplr": "workspace:*",
     "@thorswap-lib/keystore": "workspace:*",
     "@thorswap-lib/ledger": "workspace:*",
+    "@thorswap-lib/okx": "workspace:*",
     "@thorswap-lib/swapkit-api": "workspace:*",
     "@thorswap-lib/swapkit-core": "workspace:*",
     "@thorswap-lib/swapkit-entities": "workspace:*",
     "@thorswap-lib/toolbox-cosmos": "workspace:*",
     "@thorswap-lib/toolbox-evm": "workspace:*",
     "@thorswap-lib/toolbox-utxo": "workspace:*",
-    "@thorswap-lib/keepkey": "workspace:*",
     "@thorswap-lib/trezor": "workspace:*",
     "@thorswap-lib/types": "workspace:*",
     "@thorswap-lib/walletconnect": "workspace:*",
     "@thorswap-lib/xdefi": "workspace:*",
-    "@thorswap-lib/okx": "workspace:*",
     "buffer": "6.0.3",
     "crypto-browserify": "3.12.0",
+    "dotenv": "^16.3.1",
     "https-browserify": "1.0.0",
     "os-browserify": "0.3.0",
     "path-browserify": "1.0.1",
